create init folder
