+++ src/init/clean_data.c
