+++ src/init/fill_data_with_null.c
